Fix image info showed in Gallery
