+++ simplecropview/src/main/java/com/isseiaoki/simplecropview/CropImageView.java
@@ -41,7 +41,6 @@ import android.view.animation.DecelerateInterpolator;
 import android.view.animation.Interpolator;
 import android.widget.ImageView;
 
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -1334,6 +1333,7 @@ public class CropImageView extends ImageView {
             if (outputStream != null) {
                 bitmap.compress(mCompressFormat, mCompressQuality, outputStream);
                 Utils.copyExifInfo(getContext(), mSourceUri, uri, bitmap.getWidth(), bitmap.getHeight());
+                Utils.updateGalleryInfo(getContext(), uri);
             }
         } catch (IOException e) {
             Logger.e("An error occurred while saving the image: " + uri, e);
